Create README.md
"Delve into the heart of football with FIFA Prediction 2023! This repository showcases a college project that harnesses the power of data science to predict FIFA match outcomes. From preprocessing and exploratory data analysis to machine learning and AWS integration, join us on a thrilling journey through the beautiful game. Discover the winning formula and uncover the future champions of football!"